js/builtins/Object.prototype.preventExtension.html

<!doctype html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<meta charset="utf-8">
<title>Object.prototype.preventExtensions</title>
<link rel="author" title="Masaya Iseki" href="mailto:iseki.m.aa@gmail.com">
<link rel="help" href="http://www.ecma-international.org/ecma-262/5.1/#sec-15.2.3.10">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
test(function() {
  var obj = new Object();
  assert_true(Object.isExtensible(obj));
  obj.prop = 'exist';

  Object.preventExtensions(obj);
  assert_false(Object.isExtensible(obj));

  obj.extension = 'This property should not be added';
  assert_equals(undefined, obj.extension, 'Confirtm to prevent adding property.');

  obj.prop = 'changed';
  assert_equals('changed', obj.prop,
        'Confirm to be able to change a property value.');

  delete obj.prop;
  assert_equals(undefined, obj.prop, 'Confirm to be able to delete a property.');
});
</script>

</body>
</html>
